requirement.txt
PyCRS==1.0.2
leafmap==0.29.6
earthpy==0.9.4
shapely==2.0.2
ipyleaflet==0.18.0
httpx-auth==0.18.0
httpx==0.25.2
adlfs==2023.10.0
stackstac==0.5.0
xarray==2023.7.0
xpystac==0.1.3
rioxarray==0.15.0
rasterio==1.3.9
up42-py==0.33.1
